chore: rename useCreateNotice
