+++ src/features/notice/lib/useNoticeDetail.ts
@@ -3,7 +3,7 @@ import { type Notice } from '@/entities/notice/model/types';
 import { fetchNoticeById } from '@/shared/api/fetchNoticeById';
 
 export const useNoticeDetail = (id: string) => {
-  const [notice, setNotice] = useState<Notice | null>(null);
+  const [notice, setNotice] = useState<Notice>();
   const [isLoading, setIsLoading] = useState(true);
 
   useEffect(() => {
